ingest_server/requirements.txt
python_dotenv==1.0.1
flask==2.2.2
werkzeug<3
psycopg2-binary==2.9.10
alembic==1.14.0
sqlalchemy==2.0.36